% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullR.shiny.R
\name{WeibullR.shiny}
\alias{WeibullR.shiny}
\title{A Shiny Weibull Analysis App.}
\usage{
WeibullR.shiny()
}
\value{
This function does not return a value.
}
\description{
WeibullR.shiny is a Shiny web app for Weibull Analysis from WeibullR.
}
\examples{
if (interactive()) {
  WeibullR.shiny()
}
}
\seealso{
\url{https://paulgovan.github.io/WeibullR.shiny/}
}
